revert changes from path config
