+++ lstmcpipe/tests/test_workflow_management.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from ruamel.yaml import YAML
+import yaml
 import pytest
 from pathlib import Path
 
@@ -16,7 +16,7 @@ def test_save_log_to_file():
 
     assert outfile_yml.exists()
     with open(outfile_yml) as f:
-        log = YAML().load(f)
+        log = yaml.safe_load(f)
 
     assert "NoKEY" in log.keys()
     assert isinstance(log["NoKEY"], dict)
